fix Add GameRule, 게임가이드 내부 메서드 이동
