+++ RacingCar/Step3/Model/GameGuide.swift
@@ -13,34 +13,3 @@ enum GameGuide {
     static let attemptCount = "시도할 횟수는 몇 회인가요?"
 }
 
-extension GameGuide {
-    static func randomNumber() -> Int {
-        return Int.random(in: 0...9)
-    }
-    static func numberFourOrMorePlusOneOtherWisePlusZero(_ randomNumber: Int) -> Int {
-        guard randomNumber >= 4 else { return 0 }
-        return 1
-    }
-    static func splitComma(_ input: String) -> [String] {
-        return input.components(separatedBy: ",").map{String($0)}
-    }
-    
-    @discardableResult
-    static func validCheck(_ carName: String) throws -> Bool {
-        if carName.isEmpty {
-            throw InputError.emptyName
-        }
-        if carName.count > 5 {
-            throw InputError.overCountNameLength
-        }
-        return true
-    }
-    
-    static func moveCars(model cars: [Car], round: Int) -> [Car] {
-        guard round > 1 else { return cars }
-        for car in cars {
-            car.move(GameGuide.numberFourOrMorePlusOneOtherWisePlusZero(GameGuide.randomNumber()))
-        }
-        return cars
-    }
-}
